refactored URI conversion (#15)
* refactored URI conversion
