+++ doc/changes/changes_1.0.0.md
@@ -17,11 +17,14 @@ Added missing functionality to list and delete files in BucketFS and improved co
 * Updated `com.exasol:bucketfs-java:2.3.0` to `2.4.1`
 * Updated `com.exasol:error-reporting-java:0.4.1` to `1.0.0`
 * Updated `info.picocli:picocli:4.6.3` to `4.7.0`
-* Updated `nl.jqno.equalsverifier:equalsverifier:3.10` to `3.12.1`
+* Removed `nl.jqno.equalsverifier:equalsverifier:3.10`
+* Removed `org.itsallcode:junit5-system-extensions:1.2.0`
 
 ### Test Dependency Updates
 
 * Updated `com.exasol:exasol-testcontainers:6.1.1` to `6.4.0`
+* Added `nl.jqno.equalsverifier:equalsverifier:3.12.1`
+* Added `org.itsallcode:junit5-system-extensions:1.2.0`
 * Updated `org.junit.jupiter:junit-jupiter-engine:5.8.2` to `5.9.1`
 * Updated `org.junit.jupiter:junit-jupiter-params:5.8.2` to `5.9.1`
 * Updated `org.mockito:mockito-junit-jupiter:4.5.1` to `4.9.0`
