+++ src/main/java/com/exasol/bucketfs/url/OsCheck.java
@@ -1,4 +1,4 @@
-package com.exasol.bucketfs.client;
+package com.exasol.bucketfs.url;
 
 import java.util.Locale;
